+++ cmd/goatak_server/marti_api.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/kdudkov/goatak/cmd/goatak_server/mp"
 	"github.com/kdudkov/goatak/internal/pm"
+	"github.com/kdudkov/goatak/pkg/cot"
 
 	"github.com/kdudkov/goatak/pkg/cotproto"
 	"github.com/kdudkov/goatak/pkg/model"
@@ -592,12 +593,18 @@ func getVideoPostHandler(app *App) air.Handler {
 
 func getXmlHandler(app *App) air.Handler {
 	return func(req *air.Request, res *air.Response) error {
-		var evt *cotproto.CotEvent
+		uid := getStringParam(req, "uid")
 
-		if item := app.items.Get(getStringParam(req, "uid")); item != nil {
+		if uid == "" {
+			res.Status = http.StatusBadRequest
+			return res.WriteString("error")
+		}
+
+		var evt *cotproto.CotEvent
+		if item := app.items.Get(uid); item != nil {
 			evt = item.GetMsg().GetTakMessage().GetCotEvent()
 		} else {
-			di := app.missions.GetPoint(getStringParam(req, "uid"))
+			di := app.missions.GetPoint(uid)
 			if di != nil {
 				evt = di.GetEvent()
 			}
@@ -609,7 +616,7 @@ func getXmlHandler(app *App) air.Handler {
 			return nil
 		}
 
-		return res.WriteXML(evt)
+		return res.WriteXML(cot.CotToEvent(evt))
 	}
 }